GEODE-3271: Separating WanCommands into multiple command classes
