+++ geode-core/src/main/java/org/apache/geode/management/internal/web/controllers/WanCommandsController.java
@@ -15,9 +15,6 @@
 
 package org.apache.geode.management.internal.web.controllers;
 
-import org.apache.geode.internal.lang.StringUtils;
-import org.apache.geode.management.internal.cli.i18n.CliStrings;
-import org.apache.geode.management.internal.cli.util.CommandStringBuilder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -25,11 +22,28 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import org.apache.geode.internal.lang.StringUtils;
+import org.apache.geode.management.internal.cli.i18n.CliStrings;
+import org.apache.geode.management.internal.cli.util.CommandStringBuilder;
+
 /**
  * The WanCommandsController class implements the GemFire Management REST API web service endpoints
  * for the Gfsh WAN Commands.
- * 
- * @see org.apache.geode.management.internal.cli.commands.WanCommands
+ *
+ * @see org.apache.geode.management.internal.cli.commands.CreateGatewayReceiverCommand
+ * @see org.apache.geode.management.internal.cli.commands.CreateGatewaySenderCommand
+ * @see org.apache.geode.management.internal.cli.commands.DestroyGatewaySenderCommand
+ * @see org.apache.geode.management.internal.cli.commands.GatewayCommandsUtils
+ * @see org.apache.geode.management.internal.cli.commands.ListGatewayCommand
+ * @see org.apache.geode.management.internal.cli.commands.LoadBalanceGatewaySenderCommand
+ * @see org.apache.geode.management.internal.cli.commands.PauseGatewaySenderCommand
+ * @see org.apache.geode.management.internal.cli.commands.ResumeGatewaySenderCommand
+ * @see org.apache.geode.management.internal.cli.commands.StartGatewayReceiverCommand
+ * @see org.apache.geode.management.internal.cli.commands.StartGatewaySenderCommand
+ * @see org.apache.geode.management.internal.cli.commands.StatusGatewayReceiverCommand
+ * @see org.apache.geode.management.internal.cli.commands.StatusGatewaySenderCommand
+ * @see org.apache.geode.management.internal.cli.commands.StopGatewayReceiverCommand
+ * @see org.apache.geode.management.internal.cli.commands.StopGatewaySenderCommand
  * @see org.apache.geode.management.internal.web.controllers.AbstractCommandsController
  * @see org.springframework.stereotype.Controller
  * @see org.springframework.web.bind.annotation.RequestMapping
