+++ geode-wan/src/test/java/org/apache/geode/internal/cache/wan/wancommand/StartGatewayReceiverCommandDUnitTest.java
@@ -14,6 +14,18 @@
  */
 package org.apache.geode.internal.cache.wan.wancommand;
 
+import static org.apache.geode.test.dunit.Assert.assertEquals;
+import static org.apache.geode.test.dunit.Assert.assertFalse;
+import static org.apache.geode.test.dunit.Assert.assertTrue;
+import static org.apache.geode.test.dunit.Assert.fail;
+import static org.apache.geode.test.dunit.LogWriterUtils.getLogWriter;
+import static org.apache.geode.test.dunit.Wait.pause;
+
+import java.util.List;
+
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+
 import org.apache.geode.distributed.DistributedMember;
 import org.apache.geode.management.cli.Result;
 import org.apache.geode.management.internal.cli.i18n.CliStrings;
@@ -23,48 +35,27 @@ import org.apache.geode.test.dunit.Host;
 import org.apache.geode.test.dunit.VM;
 import org.apache.geode.test.junit.categories.DistributedTest;
 import org.apache.geode.test.junit.categories.FlakyTest;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-
-import java.util.List;
-import java.util.Properties;
-
-import static org.apache.geode.distributed.ConfigurationProperties.LOCATORS;
-import static org.apache.geode.distributed.ConfigurationProperties.MCAST_PORT;
-import static org.apache.geode.test.dunit.Assert.*;
-import static org.apache.geode.test.dunit.LogWriterUtils.getLogWriter;
-import static org.apache.geode.test.dunit.Wait.pause;
 
 @Category(DistributedTest.class)
-public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase {
-
-  private static final long serialVersionUID = 1L;
-
+public class StartGatewayReceiverCommandDUnitTest extends WANCommandTestBase {
   /**
    * Test wan commands for error in input 1> start gateway-sender command needs only one of member
    * or group.
    */
   @Test
   public void testStartGatewayReceiver_ErrorConditions() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createReceiver(punePort));
 
-    final DistributedMember vm1Member = vm3.invoke(() -> getMember());
-
+    final DistributedMember vm1Member = vm3.invoke(this::getMember);
     String command = CliStrings.START_GATEWAYRECEIVER + " --" + CliStrings.MEMBER + "="
         + vm1Member.getId() + " --" + CliStrings.GROUP + "=RG1";
-
     CommandResult cmdResult = executeCommand(command);
+
     if (cmdResult != null) {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter()
@@ -79,21 +70,14 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
   @Category(FlakyTest.class) // GEODE-1448
   @Test
   public void testStartGatewayReceiver() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createReceiver(punePort));
     vm4.invoke(() -> createReceiver(punePort));
     vm5.invoke(() -> createReceiver(punePort));
-
     vm3.invoke(() -> verifyReceiverState(false));
     vm4.invoke(() -> verifyReceiverState(false));
     vm5.invoke(() -> verifyReceiverState(false));
@@ -101,10 +85,10 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
     pause(10000);
     String command = CliStrings.START_GATEWAYRECEIVER;
     CommandResult cmdResult = executeCommand(command);
+
     if (cmdResult != null) {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter().info("testStartGatewayReceiver stringResult : " + strCmdResult + ">>>>");
-
       TabularResultData resultData = (TabularResultData) cmdResult.getResultData();
       List<String> status = resultData.retrieveAllValues("Result");
       assertEquals(4, status.size());
@@ -112,7 +96,6 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
     } else {
       fail("testStartGatewayReceiver failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifyReceiverState(true));
     vm4.invoke(() -> verifyReceiverState(true));
     vm5.invoke(() -> verifyReceiverState(true));
@@ -123,21 +106,14 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
    */
   @Test
   public void testStartGatewayReceiver_onMember() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createReceiver(punePort));
     vm4.invoke(() -> createReceiver(punePort));
     vm5.invoke(() -> createReceiver(punePort));
-
     vm3.invoke(() -> verifyReceiverState(false));
     vm4.invoke(() -> verifyReceiverState(false));
     vm5.invoke(() -> verifyReceiverState(false));
@@ -146,8 +122,8 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
     pause(10000);
     String command =
         CliStrings.START_GATEWAYRECEIVER + " --" + CliStrings.MEMBER + "=" + vm1Member.getId();
-
     CommandResult cmdResult = executeCommand(command);
+
     if (cmdResult != null) {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter()
@@ -157,7 +133,6 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
     } else {
       fail("testStartGatewayReceiver failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifyReceiverState(true));
     vm4.invoke(() -> verifyReceiverState(false));
     vm5.invoke(() -> verifyReceiverState(false));
@@ -168,21 +143,14 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
    */
   @Test
   public void testStartGatewayReceiver_Group() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createReceiverWithGroup(punePort, "RG1"));
     vm4.invoke(() -> createReceiverWithGroup(punePort, "RG1"));
     vm5.invoke(() -> createReceiverWithGroup(punePort, "RG1"));
-
     vm3.invoke(() -> verifyReceiverState(false));
     vm4.invoke(() -> verifyReceiverState(false));
     vm5.invoke(() -> verifyReceiverState(false));
@@ -190,11 +158,11 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
     pause(10000);
     String command = CliStrings.START_GATEWAYRECEIVER + " --" + CliStrings.GROUP + "=RG1";
     CommandResult cmdResult = executeCommand(command);
+
     if (cmdResult != null) {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter().info("testStartGatewayReceiver_Group stringResult : " + strCmdResult + ">>>>");
       assertEquals(Result.Status.OK, cmdResult.getStatus());
-
       TabularResultData resultData = (TabularResultData) cmdResult.getResultData();
       List<String> status = resultData.retrieveAllValues("Result");
       assertEquals(3, status.size());
@@ -203,7 +171,6 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
     } else {
       fail("testStartGatewayReceiver_Group failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifyReceiverState(true));
     vm4.invoke(() -> verifyReceiverState(true));
     vm5.invoke(() -> verifyReceiverState(true));
@@ -216,23 +183,16 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
    */
   @Test
   public void testStartGatewayReceiver_MultipleGroup() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createReceiverWithGroup(punePort, "RG1"));
     vm4.invoke(() -> createReceiverWithGroup(punePort, "RG1"));
     vm5.invoke(() -> createReceiverWithGroup(punePort, "RG1, RG2"));
     vm6.invoke(() -> createReceiverWithGroup(punePort, "RG1, RG2"));
     vm7.invoke(() -> createReceiverWithGroup(punePort, "RG3"));
-
     vm3.invoke(() -> verifyReceiverState(false));
     vm4.invoke(() -> verifyReceiverState(false));
     vm5.invoke(() -> verifyReceiverState(false));
@@ -242,11 +202,11 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
     pause(10000);
     String command = CliStrings.START_GATEWAYRECEIVER + " --" + CliStrings.GROUP + "=RG1,RG2";
     CommandResult cmdResult = executeCommand(command);
+
     if (cmdResult != null) {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter().info("testStartGatewayReceiver_Group stringResult : " + strCmdResult + ">>>>");
       assertEquals(Result.Status.OK, cmdResult.getStatus());
-
       TabularResultData resultData = (TabularResultData) cmdResult.getResultData();
       List<String> status = resultData.retrieveAllValues("Result");
       assertEquals(4, status.size());
@@ -255,12 +215,10 @@ public class WanCommandGatewayReceiverStartDUnitTest extends WANCommandTestBase 
     } else {
       fail("testStartGatewayReceiver failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifyReceiverState(true));
     vm4.invoke(() -> verifyReceiverState(true));
     vm5.invoke(() -> verifyReceiverState(true));
     vm6.invoke(() -> verifyReceiverState(true));
     vm7.invoke(() -> verifyReceiverState(false));
   }
-
 }
