+++ geode-wan/src/test/java/org/apache/geode/internal/cache/wan/wancommand/StopGatewayReceiverCommandDUnitTest.java
@@ -14,6 +14,18 @@
  */
 package org.apache.geode.internal.cache.wan.wancommand;
 
+import static org.apache.geode.test.dunit.Assert.assertEquals;
+import static org.apache.geode.test.dunit.Assert.assertFalse;
+import static org.apache.geode.test.dunit.Assert.assertTrue;
+import static org.apache.geode.test.dunit.Assert.fail;
+import static org.apache.geode.test.dunit.LogWriterUtils.getLogWriter;
+import static org.apache.geode.test.dunit.Wait.pause;
+
+import java.util.List;
+
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+
 import org.apache.geode.distributed.DistributedMember;
 import org.apache.geode.management.cli.Result;
 import org.apache.geode.management.internal.cli.i18n.CliStrings;
@@ -23,47 +35,25 @@ import org.apache.geode.test.dunit.Host;
 import org.apache.geode.test.dunit.VM;
 import org.apache.geode.test.junit.categories.DistributedTest;
 import org.apache.geode.test.junit.categories.FlakyTest;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-
-import java.util.List;
-import java.util.Properties;
-
-import static org.apache.geode.distributed.ConfigurationProperties.LOCATORS;
-import static org.apache.geode.distributed.ConfigurationProperties.MCAST_PORT;
-import static org.apache.geode.test.dunit.Assert.*;
-import static org.apache.geode.test.dunit.LogWriterUtils.getLogWriter;
-import static org.apache.geode.test.dunit.Wait.pause;
 
 @Category(DistributedTest.class)
-public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
-
-  private static final long serialVersionUID = 1L;
-
+public class StopGatewayReceiverCommandDUnitTest extends WANCommandTestBase {
   /**
    * Test wan commands for error in input 1> start gateway-sender command needs only one of member
    * or group.
    */
   @Test
   public void testStopGatewayReceiver_ErrorConditions() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createReceiver(punePort));
 
-    final DistributedMember vm1Member = vm3.invoke(() -> getMember());
-
+    final DistributedMember vm1Member = vm3.invoke(this::getMember);
     String command = CliStrings.STOP_GATEWAYRECEIVER + " --" + CliStrings.MEMBER + "="
         + vm1Member.getId() + " --" + CliStrings.GROUP + "=RG1";
-
     CommandResult cmdResult = executeCommand(command);
     if (cmdResult != null) {
       String strCmdResult = commandResultToString(cmdResult);
@@ -79,21 +69,14 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
   @Category(FlakyTest.class) // GEODE-1418
   @Test
   public void testStopGatewayReceiver() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createAndStartReceiver(punePort));
     vm4.invoke(() -> createAndStartReceiver(punePort));
     vm5.invoke(() -> createAndStartReceiver(punePort));
-
     vm3.invoke(() -> verifyReceiverState(true));
     vm4.invoke(() -> verifyReceiverState(true));
     vm5.invoke(() -> verifyReceiverState(true));
@@ -104,7 +87,6 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
     if (cmdResult != null) {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter().info("testStopGatewayReceiver stringResult : " + strCmdResult + ">>>>");
-
       TabularResultData resultData = (TabularResultData) cmdResult.getResultData();
       List<String> status = resultData.retrieveAllValues("Result");
       assertEquals(4, status.size());
@@ -113,7 +95,6 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
     } else {
       fail("testStopGatewayReceiver failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifyReceiverState(false));
     vm4.invoke(() -> verifyReceiverState(false));
     vm5.invoke(() -> verifyReceiverState(false));
@@ -124,30 +105,22 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
    */
   @Test
   public void testStopGatewayReceiver_onMember() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createAndStartReceiver(punePort));
     vm4.invoke(() -> createAndStartReceiver(punePort));
     vm5.invoke(() -> createAndStartReceiver(punePort));
-
     vm3.invoke(() -> verifyReceiverState(true));
     vm4.invoke(() -> verifyReceiverState(true));
     vm5.invoke(() -> verifyReceiverState(true));
 
-    final DistributedMember vm1Member = vm3.invoke(() -> getMember());
+    final DistributedMember vm1Member = vm3.invoke(this::getMember);
     pause(10000);
     String command =
         CliStrings.STOP_GATEWAYRECEIVER + " --" + CliStrings.MEMBER + "=" + vm1Member.getId();
-
     CommandResult cmdResult = executeCommand(command);
     if (cmdResult != null) {
       String strCmdResult = commandResultToString(cmdResult);
@@ -158,7 +131,6 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
     } else {
       fail("testStopGatewayReceiver failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifyReceiverState(false));
     vm4.invoke(() -> verifyReceiverState(true));
     vm5.invoke(() -> verifyReceiverState(true));
@@ -169,21 +141,14 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
    */
   @Test
   public void testStopGatewayReceiver_Group() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createAndStartReceiverWithGroup(punePort, "RG1"));
     vm4.invoke(() -> createAndStartReceiverWithGroup(punePort, "RG1"));
     vm5.invoke(() -> createAndStartReceiverWithGroup(punePort, "RG1"));
-
     vm3.invoke(() -> verifyReceiverState(true));
     vm4.invoke(() -> verifyReceiverState(true));
     vm5.invoke(() -> verifyReceiverState(true));
@@ -204,7 +169,6 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
     } else {
       fail("testStopGatewayReceiver_Group failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifyReceiverState(false));
     vm4.invoke(() -> verifyReceiverState(false));
     vm5.invoke(() -> verifyReceiverState(false));
@@ -217,23 +181,16 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
    */
   @Test
   public void testStopGatewayReceiver_MultipleGroup() {
-
     VM puneLocator = Host.getLocator();
-    int punePort = puneLocator.invoke(() -> getLocatorPort());
+    int punePort = puneLocator.invoke(this::getLocatorPort);
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createAndStartReceiverWithGroup(punePort, "RG1"));
     vm4.invoke(() -> createAndStartReceiverWithGroup(punePort, "RG1"));
     vm5.invoke(() -> createAndStartReceiverWithGroup(punePort, "RG1"));
     vm6.invoke(() -> createAndStartReceiverWithGroup(punePort, "RG1, RG2"));
     vm7.invoke(() -> createAndStartReceiverWithGroup(punePort, "RG3"));
-
     vm3.invoke(() -> verifyReceiverState(true));
     vm4.invoke(() -> verifyReceiverState(true));
     vm5.invoke(() -> verifyReceiverState(true));
@@ -247,7 +204,6 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter().info("testStopGatewayReceiver_Group stringResult : " + strCmdResult + ">>>>");
       assertEquals(Result.Status.OK, cmdResult.getStatus());
-
       TabularResultData resultData = (TabularResultData) cmdResult.getResultData();
       List<String> status = resultData.retrieveAllValues("Result");
       assertEquals(4, status.size());
@@ -256,7 +212,6 @@ public class WanCommandGatewayReceiverStopDUnitTest extends WANCommandTestBase {
     } else {
       fail("testStopGatewayReceiver failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifyReceiverState(false));
     vm4.invoke(() -> verifyReceiverState(false));
     vm5.invoke(() -> verifyReceiverState(false));
