+++ geode-wan/src/test/java/org/apache/geode/internal/cache/wan/wancommand/StopGatewaySenderCommandDUnitTest.java
@@ -14,6 +14,18 @@
  */
 package org.apache.geode.internal.cache.wan.wancommand;
 
+import static org.apache.geode.test.dunit.Assert.assertEquals;
+import static org.apache.geode.test.dunit.Assert.assertFalse;
+import static org.apache.geode.test.dunit.Assert.assertTrue;
+import static org.apache.geode.test.dunit.Assert.fail;
+import static org.apache.geode.test.dunit.LogWriterUtils.getLogWriter;
+import static org.apache.geode.test.dunit.Wait.pause;
+
+import java.util.List;
+
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+
 import org.apache.geode.distributed.DistributedMember;
 import org.apache.geode.management.cli.Result;
 import org.apache.geode.management.internal.cli.i18n.CliStrings;
@@ -21,47 +33,19 @@ import org.apache.geode.management.internal.cli.result.CommandResult;
 import org.apache.geode.management.internal.cli.result.TabularResultData;
 import org.apache.geode.test.dunit.IgnoredException;
 import org.apache.geode.test.junit.categories.DistributedTest;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-
-import java.util.List;
-import java.util.Properties;
-
-import static org.apache.geode.distributed.ConfigurationProperties.*;
-import static org.apache.geode.test.dunit.Assert.*;
-import static org.apache.geode.test.dunit.LogWriterUtils.getLogWriter;
-import static org.apache.geode.test.dunit.Wait.pause;
 
 @Category(DistributedTest.class)
-public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
-
-  private static final long serialVersionUID = 1L;
-
-  private CommandResult executeCommandWithIgnoredExceptions(String command) {
-    final IgnoredException exln = IgnoredException.addIgnoredException("Could not connect");
-    CommandResult commandResult = executeCommand(command);
-    exln.remove();
-    return commandResult;
-  }
-
+public class StopGatewaySenderCommandDUnitTest extends WANCommandTestBase {
   @Test
   public void testStopGatewaySender_ErrorConditions() {
-
     Integer punePort = vm1.invoke(() -> createFirstLocatorWithDSId(1));
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(DISTRIBUTED_SYSTEM_ID, "1");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createCache(punePort));
     vm3.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
 
-    final DistributedMember vm1Member = vm3.invoke(() -> getMember());
-
+    final DistributedMember vm1Member = vm3.invoke(this::getMember);
     String command = CliStrings.STOP_GATEWAYSENDER + " --" + CliStrings.STOP_GATEWAYSENDER__ID
         + "=ln --" + CliStrings.MEMBER + "=" + vm1Member.getId() + " --" + CliStrings.GROUP
         + "=SenderGroup1";
@@ -78,28 +62,19 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
 
   @Test
   public void testStopGatewaySender() {
-
     Integer punePort = vm1.invoke(() -> createFirstLocatorWithDSId(1));
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(DISTRIBUTED_SYSTEM_ID, "1");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createCache(punePort));
     vm3.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
     vm4.invoke(() -> createCache(punePort));
     vm4.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
     vm5.invoke(() -> createCache(punePort));
     vm5.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
-
     vm3.invoke(() -> startSender("ln"));
     vm4.invoke(() -> startSender("ln"));
     vm5.invoke(() -> startSender("ln"));
-
     vm3.invoke(() -> verifySenderState("ln", true, false));
     vm4.invoke(() -> verifySenderState("ln", true, false));
     vm5.invoke(() -> verifySenderState("ln", true, false));
@@ -112,7 +87,6 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter().info("testStopGatewaySender stringResult : " + strCmdResult + ">>>>");
       assertEquals(Result.Status.OK, cmdResult.getStatus());
-
       TabularResultData resultData = (TabularResultData) cmdResult.getResultData();
       List<String> status = resultData.retrieveAllValues("Result");
       assertEquals(5, status.size());
@@ -121,7 +95,6 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
     } else {
       fail("testStopGatewaySender failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifySenderState("ln", false, false));
     vm4.invoke(() -> verifySenderState("ln", false, false));
     vm5.invoke(() -> verifySenderState("ln", false, false));
@@ -132,25 +105,16 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
    */
   @Test
   public void testStopGatewaySender_onMember() {
-
     Integer punePort = vm1.invoke(() -> createFirstLocatorWithDSId(1));
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(DISTRIBUTED_SYSTEM_ID, "1");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createCache(punePort));
     vm3.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
-
     vm3.invoke(() -> startSender("ln"));
-
     vm3.invoke(() -> verifySenderState("ln", true, false));
 
-    final DistributedMember vm1Member = vm3.invoke(() -> getMember());
+    final DistributedMember vm1Member = vm3.invoke(this::getMember);
     pause(10000);
     String command = CliStrings.STOP_GATEWAYSENDER + " --" + CliStrings.STOP_GATEWAYSENDER__ID
         + "=ln --" + CliStrings.MEMBER + "=" + vm1Member.getId();
@@ -163,7 +127,6 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
     } else {
       fail("testStopGatewaySender failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifySenderState("ln", false, false));
   }
 
@@ -172,28 +135,19 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
    */
   @Test
   public void testStopGatewaySender_Group() {
-
     Integer punePort = vm1.invoke(() -> createFirstLocatorWithDSId(1));
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(DISTRIBUTED_SYSTEM_ID, "1");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createCacheWithGroups(punePort, "SenderGroup1"));
     vm3.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
     vm4.invoke(() -> createCacheWithGroups(punePort, "SenderGroup1"));
     vm4.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
     vm5.invoke(() -> createCacheWithGroups(punePort, "SenderGroup1"));
     vm5.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
-
     vm3.invoke(() -> startSender("ln"));
     vm4.invoke(() -> startSender("ln"));
     vm5.invoke(() -> startSender("ln"));
-
     vm3.invoke(() -> verifySenderState("ln", true, false));
     vm4.invoke(() -> verifySenderState("ln", true, false));
     vm5.invoke(() -> verifySenderState("ln", true, false));
@@ -206,7 +160,6 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
       String strCmdResult = commandResultToString(cmdResult);
       getLogWriter().info("testStopGatewaySender_Group stringResult : " + strCmdResult + ">>>>");
       assertEquals(Result.Status.OK, cmdResult.getStatus());
-
       TabularResultData resultData = (TabularResultData) cmdResult.getResultData();
       List<String> status = resultData.retrieveAllValues("Result");
       assertEquals(3, status.size());
@@ -215,7 +168,6 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
     } else {
       fail("testStopGatewaySender failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifySenderState("ln", false, false));
     vm4.invoke(() -> verifySenderState("ln", false, false));
     vm5.invoke(() -> verifySenderState("ln", false, false));
@@ -227,17 +179,10 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
    */
   @Test
   public void testStopGatewaySender_MultipleGroup() {
-
     Integer punePort = vm1.invoke(() -> createFirstLocatorWithDSId(1));
+    propsSetUp(punePort);
 
-    Properties props = getDistributedSystemProperties();
-    props.setProperty(MCAST_PORT, "0");
-    props.setProperty(DISTRIBUTED_SYSTEM_ID, "1");
-    props.setProperty(LOCATORS, "localhost[" + punePort + "]");
-    setUpJmxManagerOnVm0ThenConnect(props);
-
-    Integer nyPort = vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
-
+    vm2.invoke(() -> createFirstRemoteLocator(2, punePort));
     vm3.invoke(() -> createCacheWithGroups(punePort, "SenderGroup1"));
     vm3.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
     vm4.invoke(() -> createCacheWithGroups(punePort, "SenderGroup1"));
@@ -248,13 +193,11 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
     vm6.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
     vm7.invoke(() -> createCacheWithGroups(punePort, "SenderGroup3"));
     vm7.invoke(() -> createSender("ln", 2, false, 100, 400, false, false, null, true));
-
     vm3.invoke(() -> startSender("ln"));
     vm4.invoke(() -> startSender("ln"));
     vm5.invoke(() -> startSender("ln"));
     vm6.invoke(() -> startSender("ln"));
     vm7.invoke(() -> startSender("ln"));
-
     vm3.invoke(() -> verifySenderState("ln", true, false));
     vm4.invoke(() -> verifySenderState("ln", true, false));
     vm5.invoke(() -> verifySenderState("ln", true, false));
@@ -277,11 +220,17 @@ public class WanCommandGatewaySenderStopDUnitTest extends WANCommandTestBase {
     } else {
       fail("testStopGatewaySender failed as did not get CommandResult");
     }
-
     vm3.invoke(() -> verifySenderState("ln", false, false));
     vm4.invoke(() -> verifySenderState("ln", false, false));
     vm5.invoke(() -> verifySenderState("ln", false, false));
     vm6.invoke(() -> verifySenderState("ln", false, false));
     vm7.invoke(() -> verifySenderState("ln", true, false));
+  }
+
+  private CommandResult executeCommandWithIgnoredExceptions(String command) {
+    final IgnoredException exln = IgnoredException.addIgnoredException("Could not connect");
+    CommandResult commandResult = executeCommand(command);
+    exln.remove();
+    return commandResult;
   }
 }
